tests refactored except for mock integration test
